Clean trailing dash
Usernames that end in _ are cleaned to end in -, but that makes a Rancher service name with -- which Rancher 1.x doesn't like.  This change replaces a trailing - in the cleaned userid with -0.  So user_name_ would before produce a Rancher service name like 'user-name--hash' but now is like 'user-name-0-hash'.